infra/scenarios/azure_functions_flex_consumption/src/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
#, and azure-functions-worker doesn't exist on pypi.
azure-functions
